refactored minChipSize from int to float

+++ src/main/java/com/arthur/blackjack/config/GameSettings.java
@@ -3,12 +3,12 @@ package com.arthur.blackjack.config;
 public interface GameSettings {
     float getBankroll();
     int getBetSize();
-    int getMinChipSize();
+    float getMinChipSize();
     int getBetSpread();
     int getMaxRounds();
     void setBankroll(float bankroll);
     void setBetSize(int betSize);
-    void setMinChipSize(int minChipSize);
+    void setMinChipSize(float minChipSize);
     void setBetSpread(int betSpread);
     void setMaxRounds(int maxRounds);
 }
